Enum scheduler::HostIter

pub enum HostIter<'a, 'b, HostType: Host> {
    ThreadPerHost(&'a mut HostIter<HostType>),
    ThreadPerCore(&'a mut HostIter<'b, HostType>),
}

Supports iterating over all hosts assigned to this thread.

Variants

ThreadPerHost(&'a mut HostIter<HostType>)

ThreadPerCore(&'a mut HostIter<'b, HostType>)

Implementations

impl<'a, 'b, HostType: Host> HostIter<'a, 'b, HostType>

pub fn for_each<F>(&mut self, f: F)
where F: FnMut(HostType) -> HostType,

For each Host, calls f with the host. The Host must be returned by the closure. The ownership of the Host is transferred in and out of the closure rather than using a mutable reference since Shadow needs to put the host in a global with 'static lifetime (the worker).