Crate scheduler

Offers the ability to control or configure the active task scheduling policy.

What is and isn’t in this crate?

This crate also defines the timer interrupt handler used for preemptive task switching on each CPU. In init(), it registers that handler with the [interrupts] subsystem.

The actual task switching logic is implemented in the [task] crate. This crate re-exports that main schedule() function for convenience, legacy compatbility, and to act as an easy landing page for code search. That means that a caller need only depend on [task], not this crate, to invoke the scheduler (yield the CPU) to switch to another task.

Functions

get_priority

Returns the priority of a given task. This function returns None when a scheduler without priority is loaded.

inherit_priority

init

Initializes the scheduler on this system using the policy set at compiler time.

schedule

A re-export of task::schedule() for convenience and legacy compatibility. Yields the current CPU by selecting a new Task to run next, and then switches to that new Task.

set_priority

Changes the priority of the given task with the given priority level. Priority values must be between 40 (maximum priority) and 0 (minimum prriority). This function returns an error when a scheduler without priority is loaded.